Atualização

+++ modulo2/apis-labenusers/apis-labenusers/src/components/ListUsers/index.js
@@ -1,55 +1,76 @@
 import React from "react";
 import axios from "axios";
 
-
-
 class ListUsers extends React.Component {
     state = {
-        users: this.props.users,
+        usuarios: [],
         name: ""
     }
 
-    deleteUser = (id) => {
+    componentDidMount = () => {
+        this.getAllUsers();
+    }
+    
+    getAllUsers = () => {
         const headers = this.props.headers;
 
-        const url = `https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users/${id}`;
-
-        axios.delete(url, headers)
+        const url = "https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users"
+    
+        axios.get(url, headers)
         .then(res => {
-            alert("Usuário deletado com sucesso!");
-            this.props.getAllUsers();
+          this.setState({usuarios: res.data});
         })
-        .catch(err => {
-            alert("Ocorreu um erro, tente novamente.")
+        .catch(error => {
+          alert(error.res.data);
         })
+    }
+
+    deleteUser = (id) => {
+        if(window.confirm("Tem certeza que deseja apagar o usuário?")){
+            const headers = this.props.headers;
+
+            const url = `https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users/${id}`;
+
+            axios.delete(url, headers)
+            .then(res => {
+                alert("Usuário deletado com sucesso!");
+                this.getAllUsers();
+            })
+            .catch(err => {
+                alert("Ocorreu um erro, tente novamente.")
+            })
+        }
+        return;
     }
 
     searchUser = async() => {
         const headers = this.props.headers;
 
         const url = `https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users/search?name=${this.state.name}`;
-
+        
         try{
             const res = await axios.get(url,headers)
-            
-            this.setState({users: res.data});
-                    
-            this.props.getAllUsers();
+
+            this.setState({usuarios: res.data});
 
             this.setState({name: ""});
         }
         catch(err) {
             alert(err.response.data.message)
         }
+
+        if(this.state.usuarios.length === 0){
+            this.getAllUsers();
+        }
     }
 
     onChangeName = (event) => this.setState({name: event.target.value});
 
     render() {
-        const list = this.state.users.map(users => {
+        const list = this.state.usuarios.map(users => {
             return (
                 <li key={users.id}>
-                    {users.name}
+                    <p>{users.name}</p>
                     <button onClick={() => this.deleteUser(users.id)}>X</button>
                 </li>
             )
@@ -57,6 +78,7 @@ class ListUsers extends React.Component {
 
         return(
             <div>
+                <button onClick={this.props.goToPostUsers}>Trocar Tela</button> <br/>
                 <ul>
                     {list}
                 </ul>
